Replace frightening fromBuilder method with new fromCollection
The existing fromBuilder method reused the same builder for multiple `invert`
calls, which is not safe under multithreading: one thread might, for example,
clear the builder while another thread is adding elements.

This patch changes the implementation to get a new builder every time from the
CanBuildFrom, and puts that behind the existing fromCollection interface. The
existing injections for specific collections should not change behaviour.